Add League of Legends API project, Analyzing LOL matches notebook, and database files

+++ data_science/database_facul/atv_sql1.sql
@@ -21,7 +21,7 @@ USE `eshop` ;
 -- Table `eshop`.`cliente_rel`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `eshop`.`cliente_rel` (
-  `codcli` INT(15) NOT NULL,
+  `codcli` NUMERIC(15) NOT NULL,
   `nome` VARCHAR(100) NOT NULL,
   `rua` VARCHAR(100) NOT NULL,
   `cidade` VARCHAR(100) NOT NULL,
@@ -42,9 +42,9 @@ ENGINE = InnoDB;
 -- Table `eshop`.`cliente_vip_rel`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `eshop`.`cliente_vip_rel` (
-  `cliente_rel_codcli` INT(15) NOT NULL,
+  `cliente_rel_codcli` NUMERIC(15) NOT NULL,
   `pontos_bonificacao` INT NULL,
-  `desconto_padrao` INT(5) NULL,
+  `desconto_padrao` NUMERIC(5) NULL,
   PRIMARY KEY (`cliente_rel_codcli`),
   INDEX `fk_cliente_vip_rel_cliente_rel_idx` (`cliente_rel_codcli` ASC) VISIBLE,
   UNIQUE INDEX `cliente_rel_codcli_UNIQUE` (`cliente_rel_codcli` ASC) VISIBLE,
@@ -60,8 +60,8 @@ ENGINE = InnoDB;
 -- Table `eshop`.`cliente_especial_rel`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `eshop`.`cliente_especial_rel` (
-  `cliente_rel_codcli` INT(15) NOT NULL,
-  `desconto_padrao` INT(5) NULL,
+  `cliente_rel_codcli` NUMERIC(15) NOT NULL,
+  `desconto_padrao` NUMERIC(5) NULL,
   PRIMARY KEY (`cliente_rel_codcli`),
   CONSTRAINT `fk_cliente_especial_rel_cliente_rel1`
     FOREIGN KEY (`cliente_rel_codcli`)
@@ -75,14 +75,14 @@ ENGINE = InnoDB;
 -- Table `eshop`.`pedido_rel`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `eshop`.`pedido_rel` (
-  `codped` INT(20) NOT NULL,
+  `codped` NUMERIC(20) NOT NULL,
   `data_pedido` DATE NOT NULL,
   `data_entrega` DATE NULL,
   `rua` VARCHAR(100) NOT NULL,
   `cidade` VARCHAR(100) NOT NULL,
   `estado` CHAR(2) NULL,
   `cep` CHAR(10) NULL,
-  `cliente_rel_codcli` INT(15) NOT NULL,
+  `cliente_rel_codcli` NUMERIC(15) NOT NULL,
   PRIMARY KEY (`codped`),
   INDEX `fk_pedido_rel_cliente_rel1_idx` (`cliente_rel_codcli` ASC) VISIBLE,
   CONSTRAINT `fk_pedido_rel_cliente_rel1`
@@ -97,9 +97,9 @@ ENGINE = InnoDB;
 -- Table `eshop`.`mercadoria_rel`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `eshop`.`mercadoria_rel` (
-  `codmer` INT NOT NULL,
-  `preço` FLOAT(15,2) NOT NULL,
-  `icms` FLOAT(5,2) NOT NULL,
+  `codmer` NUMERIC NOT NULL,
+  `preço` NUMERIC(15,2) NOT NULL,
+  `icms` NUMERIC(5,2) NOT NULL,
   PRIMARY KEY (`codmer`),
   UNIQUE INDEX `codmer_UNIQUE` (`codmer` ASC) VISIBLE)
 ENGINE = InnoDB;
@@ -109,9 +109,9 @@ ENGINE = InnoDB;
 -- Table `eshop`.`item_pedido_rel`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `eshop`.`item_pedido_rel` (
-  `no_item` INT(5) NOT NULL,
-  `pedido_rel_codped` INT(20) NOT NULL,
-  `quantidade` INT(15) NOT NULL,
+  `no_item` NUMERIC(5) NOT NULL,
+  `pedido_rel_codped` NUMERIC(20) NOT NULL,
+  `quantidade` NUMERIC(15) NOT NULL,
   `desconto` FLOAT(5,2) NULL,
   `mercadoria_rel_codmer` INT NOT NULL,
   PRIMARY KEY (`no_item`, `pedido_rel_codped`),
